// src/frontend/webapp/modilist/src/services/swagger/api/models/fit-preferences-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ModilistAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface FitPreferencesDTO
 */
export interface FitPreferencesDTO {
    /**
     * 
     * @type {string}
     * @memberof FitPreferencesDTO
     */
    waistHeight?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FitPreferencesDTO
     */
    upperFit?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FitPreferencesDTO
     */
    lowerFit?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FitPreferencesDTO
     */
    legFit?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FitPreferencesDTO
     */
    skirtDressLength?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FitPreferencesDTO
     */
    shortsLength?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FitPreferencesDTO
     */
    footType?: string | null;
}